Construct a Dictionary Tree from JSON Files

This function reads JSON files from a specified directory, creates nodes for each file, and builds a hierarchical tree based on the relationships defined within the files.

Usage

construct_dictionary_tree(path = "./data/test_dictionary")

Arguments

path

The path to the directory containing JSON files.

Value

A Node object representing the root of the dictionary tree.

Examples

path <- get_path_to_example_dictionary()
print(path)
#> [1] "/home/runner/work/_temp/Library/cidtree/extdata/test_dictionary"
dd <- construct_dictionary_tree(path)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
# Print output so that it does not wrap on the screen
output <- capture.output(print(dd, 'concept_str', 'concept_type','pathString'))
cat(output, sep="\n")
#>                    levelName                                        concept_str
#> 1  0                         DICT: connect4cancer                              
#> 2   ¦--129084651              PRI_SRC: Survey                                  
#> 3   ¦   °--965707586           SEC_SRC: Smoking, Alcohol, and Sun Exposure     
#> 4   ¦       ¦--194557808          QUEST: SrvSAS_AgeCigsUseF_v1r0               
#> 5   ¦       ¦--790436165        SRC: Age first used cigarettes on a regular ...
#> 6   ¦       ¦   ¦--408696162      QUEST: SrvSAS_AgeCigsUseRegF_v1r0            
#> 7   ¦       ¦   °--965482317      QUEST: SrvSAS_NeverSmokedReg_v1r0            
#> 8   ¦       ¦--947205597        SRC: Have you ever used any of these tobacco...
#> 9   ¦       ¦   ¦--535003378      QUEST: SrvSAS_NoTobacco_v1r0                 
#> 10  ¦       ¦   ¦--686310465      QUEST: SrvSAS_TobaccoPipe_v1r0               
#> 11  ¦       ¦   °--712653855      QUEST: SrvSAS_Cigarettes_v1r0                
#> 12  ¦       ¦--639684251          QUEST: SrvSAS_UseCigsNow_v1r0                
#> 13  ¦       ¦--763164658          QUEST: SrvSAS_CigarettesLife_v1r0            
#> 14  ¦       °--798549704          QUEST: SrvSAS_UseCigsLast_v1r0               
#> 15  ¦--192505768              PRI_SRC: Recruitment                             
#> 16  ¦   °--214456996           SEC_SRC: Eligibility Screener                   
#> 17  ¦       ¦--142654897        SRC: How did you hear about this study? (Sel...
#> 18  ¦       ¦   ¦--461488577      QUEST: RcrtES_Aware_v1r0_Email               
#> 19  ¦       ¦   ¦--607081902      QUEST: RcrtES_Aware_v1r0_Invite              
#> 20  ¦       ¦   °--942255248      QUEST: RcrtES_Aware_v1r0_Mail                
#> 21  ¦       ¦--827220437          QUEST: RcrtES_Site_v1r0                      
#> 22  ¦       °--948195369          QUEST: RcrtES_PINmatch_v1r0                  
#> 23  ¦--104430631              RESP: No                                         
#> 24  ¦--125001209              RESP: Kaiser Permanente Colorado                 
#> 25  ¦--132232896              RESP: 100 or more                                
#> 26  ¦--151488193              RESP: 10 or less                                 
#> 27  ¦--181769837              RESP: Other                                      
#> 28  ¦--299561721              RESP: Yes, but rarely                            
#> 29  ¦--300267574              RESP: Kaiser Permanente Hawaii                   
#> 30  ¦--303349821              RESP: Marshfield Clinic Health System            
#> 31  ¦--317567178              RESP: In the past month                          
#> 32  ¦--327912200              RESP: Kaiser Permanente Georgia                  
#> 33  ¦--353358909              RESP: Yes                                        
#> 34  ¦--419415087              RESP: No, not at all                             
#> 35  ¦--452412599              RESP: Kaiser Permanente Northwest                
#> 36  ¦--484055234              RESP: More than a month ago, but in the past year
#> 37  ¦--486319890              RESP: 50 to 99                                   
#> 38  ¦--517700004              RESP: National Cancer Institute                  
#> 39  ¦--531629870              RESP: HealthPartners                             
#> 40  ¦--548392715              RESP: Henry Ford Health System                   
#> 41  ¦--648960871              RESP: Never                                      
#> 42  ¦--657167265              RESP: Sanford Health                             
#> 43  ¦--716761013              RESP: Yes, but some days                         
#> 44  ¦--802197176              RESP: More than 1 year ago                       
#> 45  ¦--804785430              RESP: Yes, every day                             
#> 46  ¦--805449318              RESP: 11 to 49                                   
#> 47  ¦--809703864              RESP: University of Chicago Medicine             
#> 48  °--812620303              RESP: N/A                                        
#>    concept_type                                pathString
#> 1          ROOT                                         0
#> 2       PRIMARY                               0/129084651
#> 3     SECONDARY                     0/129084651/965707586
#> 4      QUESTION           0/129084651/965707586/194557808
#> 5        SOURCE           0/129084651/965707586/790436165
#> 6      QUESTION 0/129084651/965707586/790436165/408696162
#> 7      QUESTION 0/129084651/965707586/790436165/965482317
#> 8        SOURCE           0/129084651/965707586/947205597
#> 9      QUESTION 0/129084651/965707586/947205597/535003378
#> 10     QUESTION 0/129084651/965707586/947205597/686310465
#> 11     QUESTION 0/129084651/965707586/947205597/712653855
#> 12     QUESTION           0/129084651/965707586/639684251
#> 13     QUESTION           0/129084651/965707586/763164658
#> 14     QUESTION           0/129084651/965707586/798549704
#> 15      PRIMARY                               0/192505768
#> 16    SECONDARY                     0/192505768/214456996
#> 17       SOURCE           0/192505768/214456996/142654897
#> 18     QUESTION 0/192505768/214456996/142654897/461488577
#> 19     QUESTION 0/192505768/214456996/142654897/607081902
#> 20     QUESTION 0/192505768/214456996/142654897/942255248
#> 21     QUESTION           0/192505768/214456996/827220437
#> 22     QUESTION           0/192505768/214456996/948195369
#> 23     RESPONSE                               0/104430631
#> 24     RESPONSE                               0/125001209
#> 25     RESPONSE                               0/132232896
#> 26     RESPONSE                               0/151488193
#> 27     RESPONSE                               0/181769837
#> 28     RESPONSE                               0/299561721
#> 29     RESPONSE                               0/300267574
#> 30     RESPONSE                               0/303349821
#> 31     RESPONSE                               0/317567178
#> 32     RESPONSE                               0/327912200
#> 33     RESPONSE                               0/353358909
#> 34     RESPONSE                               0/419415087
#> 35     RESPONSE                               0/452412599
#> 36     RESPONSE                               0/484055234
#> 37     RESPONSE                               0/486319890
#> 38     RESPONSE                               0/517700004
#> 39     RESPONSE                               0/531629870
#> 40     RESPONSE                               0/548392715
#> 41     RESPONSE                               0/648960871
#> 42     RESPONSE                               0/657167265
#> 43     RESPONSE                               0/716761013
#> 44     RESPONSE                               0/802197176
#> 45     RESPONSE                               0/804785430
#> 46     RESPONSE                               0/805449318
#> 47     RESPONSE                               0/809703864
#> 48     RESPONSE                               0/812620303